backend/requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
flask-socketio>=5.3.6
pymavlink>=2.4.37
pyserial>=3.5
opencv-python-headless>=4.8.0
eventlet>=0.35.1
